Main layout locked in
Container of 100% width & height contains two columns, aside and main,
that can overflow and be individually scrolled.

Aside contains the list of users.

Main contains message list and form. The chat form is "fixed" at the
bottom of the viewport and covers part of the messages area. The message
list is pushed to the bottom of its container (using flex) so that
messages are displayed near the form.